+++ templates/AzureFunctionRest/PSMDInvoke.ps1
@@ -1,0 +1,5 @@
+﻿param (
+	$Path
+)
+
+New-PSMDTemplate -FilePath "$PSScriptRoot\þnameþ.ps1" -TemplateName AzureFunctionRest -OutPath $Path -Description "Template for an Azure Function with Rest Trigger" -Author "Friedrich Weinmann" -Tags 'function', 'file', 'azure'